+++ src/components/model/article/ArticleSWRContainer/ArticleSkeltonList/ArticleSkelton/index.stories.tsx
@@ -1,0 +1,16 @@
+import type { ComponentMeta, ComponentStory } from "@storybook/react";
+
+import { ArticleSkelton } from ".";
+
+export default {
+  component: ArticleSkelton,
+  title: "model/article/ArticleSWRContainer/ArticleSkeltonList/ArticleSkelton",
+} as ComponentMeta<typeof ArticleSkelton>;
+
+const Template: ComponentStory<typeof ArticleSkelton> = (args) => (
+  <div className="h-96 w-96">
+    <ArticleSkelton {...args} />
+  </div>
+);
+
+export const Default = Template.bind({});
