+++ src/components/model/article/carousel/CarouselContainer/index.tsx
