+++ src/components/model/article/carousel/CarouselCore.tsx/index.tsx
