--- src/components/organisms/article/carousel/PopularArticleCarousel/index.spec.tsx
@@ -1,13 +1,0 @@
-import { render, screen } from "jest/test-utils";
-import { mockPopularArticles } from "mocks/data";
-
-import { PopularArticleCarousel } from ".";
-
-describe("components/organisms/PopularArticleCarousel", () => {
-  it("OK: 初期表示が正しい", () => {
-    render(<PopularArticleCarousel articles={mockPopularArticles} />);
-
-    const articleEls = screen.getAllByRole("article");
-    expect(articleEls.length).toBe(mockPopularArticles.length);
-  });
-});
